+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to this project will be documented in this file.
 - `mam` mutation `dropCachedDbInfo` to reset cached indexes.
  
 ### Fix
-- Update indexes retries if index creation has failed on the timeout. 
+- Update indexes retries if index creation has failed on the timeout.
+- Basic Debian image updated from "Stretch" to "Buster" to meet dependencies.
 
 ## 0.27.7 – Jul 31, 2020